requirements.txt
absl-py==0.9.0
astor==0.8.0
backcall==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.6.0
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
gast==0.3.3
grpcio==1.27.2
h5py==2.8.0
idna==2.9
imutils==0.5.3
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.2.0
numpy==1.16.4
olefile==0.46
parso==0.7.0
Paste==3.2.2
pickleshare==0.7.5
Pillow==7.1.2
portend==2.5
prompt-toolkit==3.0.4
protobuf==3.11.4
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pyzmq==18.1.1
repoze.lru==0.7
requests==2.23.0
Routes==2.4.1
scikit-learn==0.23.0
scipy==1.4.1
selenium==3.141.0
simplejson==3.17.0
six==1.14.0
tabulate==0.8.7
tempora==3.0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.0.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zc.lockfile==2.0
